# api/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "fields": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "propertyNames": {
              "type": "string"
            },
            "type": [
              "object",
              "null"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ApiErrorCode": {
        "description": "Closed, stable machine-readable failure codes shared by every public HTTP\nresponse. Endpoint response declarations constrain which subset is\napplicable, while the runtime wire type cannot emit an unregistered code.",
        "enum": [
          "request_forbidden",
          "unsupported_media_type",
          "payload_too_large",
          "malformed_json",
          "invalid_request",
          "resource_not_found",
          "method_not_allowed",
          "revision_conflict",
          "command_not_found",
          "command_state_conflict",
          "profile_switch_conflict",
          "serial_connection_failed",
          "camera_unavailable",
          "screenshot_unavailable",
          "destination_conflict",
          "notification_unsupported",
          "dynamic_config_unavailable",
          "unsupported_platform",
          "persistence_failed",
          "backend_unavailable",
          "internal_error"
        ],
        "type": "string"
      },
...
      "CameraSelector": {
        "oneOf": [
          {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          {
            "type": "string"
          }
        ]
      },
...
      "CommandDisplayCommand": {
        "additionalProperties": false,
        "properties": {
          "command": {
            "$ref": "#/components/schemas/CommandInfo"
          }
        },
        "required": [
          "command"
        ],
        "type": "object"
      },
      "CommandDisplayItem": {
        "discriminator": {
          "propertyName": "kind"
        },
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/CommandDisplayCommand"
              },
              {
                "additionalProperties": false,
                "properties": {
                  "kind": {
                    "enum": [
                      "command"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "kind"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/CommandDisplaySeparator"
              },
              {
                "additionalProperties": false,
                "properties": {
                  "kind": {
                    "enum": [
                      "separator"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "kind"
                ],
                "type": "object"
              }
            ]
          }
        ]
      },
      "CommandDisplaySeparator": {
        "additionalProperties": false,
        "properties": {
          "label": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "CommandInfo": {
        "additionalProperties": false,
        "properties": {
          "class_name": {
            "type": "string"
          },
          "module_path": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "name",
          "module_path",
          "class_name",
          "tags"
        ],
        "type": "object"
      },
...
      "CommandState": {
        "enum": [
          "running",
          "paused",
          "stopped",
          "error"
        ],
        "type": "string"
      },
      "DecimalString": {
        "description": "An exact non-negative integer wire value that never depends on JavaScript\n`Number` precision.",
        "example": "42",
        "pattern": "^(0|[1-9][0-9]*)$",
        "type": "string"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Common secret-safe error envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "SettingsPatchRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DecimalString"
              }
            ]
          },
          "values": {
            "$ref": "#/components/schemas/SettingsWriteValues"
          }
        },
        "required": [
          "values"
        ],
        "type": "object"
      },
...
      "Success_SettingsSnapshot": {
        "additionalProperties": false,
        "description": "Common successful JSON envelope.",
        "properties": {
          "data": {
            "additionalProperties": false,
            "properties": {
              "apply_failures": {
                "additionalProperties": {
                  "type": "string"
                },
                "propertyNames": {
                  "type": "string"
                },
                "type": "object"
              },
              "pending_restart_values": {
                "$ref": "#/components/schemas/SettingsWriteValues"
              },
              "restart_required": {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              "revision": {
                "$ref": "#/components/schemas/DecimalString"
              },
              "values": {
                "$ref": "#/components/schemas/SettingsReadValues"
              }
            },
            "required": [
              "revision",
              "values",
              "pending_restart_values",
              "restart_required",
              "apply_failures"
            ],
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "Success_StateSnapshot": {
        "additionalProperties": false,
        "description": "Common successful JSON envelope.",
        "properties": {
          "data": {
            "additionalProperties": false,
            "properties": {
              "active_profile": {
                "type": "string"
              },
              "available_profiles": {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              "camera_device": {
                "$ref": "#/components/schemas/CameraSelector"
              },
              "camera_fps": {
                "format": "double",
                "type": "number"
              },
              "camera_opened": {
                "type": "boolean"
              },
              "camera_resolution": {
                "type": "string"
              },
              "command_candidates": {
                "items": {
                  "$ref": "#/components/schemas/CommandInfo"
                },
                "type": "array"
              },
              "command_display_cache_loading": {
                "type": "boolean"
              },
              "command_display_lists": {
                "additionalProperties": {
                  "items": {
                    "$ref": "#/components/schemas/CommandDisplayItem"
                  },
                  "type": "array"
                },
                "propertyNames": {
                  "type": "string"
                },
                "type": "object"
              },
              "command_state": {
                "$ref": "#/components/schemas/CommandState"
              },
              "current_command": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "holding_buttons": {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              "is_running": {
                "type": "boolean"
              },
              "last_input": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "pending_profile": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "pid": {
                "format": "int32",
                "minimum": 0,
                "type": "integer"
              },
              "revision": {
                "$ref": "#/components/schemas/DecimalString"
              },
              "serial_baud_rate": {
                "format": "int32",
                "minimum": 0,
                "type": "integer"
              },
              "serial_connected": {
                "type": "boolean"
              },
              "serial_port": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "tags": {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            },
            "required": [
              "revision",
              "serial_baud_rate",
              "serial_connected",
              "camera_opened",
              "camera_fps",
              "camera_resolution",
              "camera_device",
              "is_running",
              "command_state",
              "command_candidates",
              "tags",
              "active_profile",
              "available_profiles",
              "holding_buttons",
              "pid",
              "command_display_lists",
              "command_display_cache_loading"
            ],
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Closed REST and WebSocket contracts for PokeCon",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "PokeCon API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/settings": {
      "get": {
        "operationId": "get_settings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_SettingsSnapshot"
                }
              }
            },
            "description": "Current canonical settings"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Settings snapshot failed"
          }
        },
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "patch_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_SettingsSnapshot"
                }
              }
            },
            "description": "Committed settings snapshot"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Malformed JSON"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Revision or runtime conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid setting transaction"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Persistence failure"
          }
        },
        "tags": [
          "settings"
        ]
      }
    },
    "/api/state": {
      "get": {
        "operationId": "get_state",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_StateSnapshot"
                }
              }
            },
            "description": "Current runtime state"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "State snapshot failed"
          }
        },
        "tags": [
          "state"
        ]
      }
    },
...
    "/ws": {
      "get": {
        "operationId": "websocket",
        "responses": {
          "101": {
            "description": "WebSocket upgrade"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Host or Origin rejected"
          }
        },
        "tags": [
          "websocket"
        ]
      }
    }
  },
...
}